Shows dynamic list bug
The items in an unordered list are added dynamically as the button is
pressed. Each list item contains an ‘input’ element. If only on item is
added to the list, and the input is typed in, the whole list item will
be destroyed. If two or more list items are added to the list, then
typing in any of the input items works fine. What gives?